Add files via upload
Website Description:

Discover the enchanting wonders of Lebanon's nature and explore the countless picturesque destinations that await you on CedarTours. Our captivating website offers a seamless and user-friendly interface, allowing you to immerse yourself in the breathtaking beauty of Lebanon's landscapes and diverse ecosystems. Prepare to be amazed as you embark on a visual journey through the natural wonders that make Lebanon truly remarkable.

Developers:
-Aya Moussawi-109145
(Please note that Aya is a first-year student, but she has provided us with great consultancy and assistance)
-Lara Ibrahim-108740
-Mariam Saab-108926
-Ali Fakhereddine-108247